+++ CHANGELOG.md
@@ -36,6 +36,7 @@
   [HexaPDF::Type::AcroForm::Field#full_field_name]
 * Moved `HexaPDF::Content::Canvas#color_space_for_components` to class method on
   [HexaPDF::Content::ColorSpace]
+* Added bit unsetter method to[HexaPDF::Utils::BitField]
 
 ### Fixed